Bug fixes:
Peer tests now use mocknet

Fixed bug caused by expired request timeouts and another in removeExpiredObjects

Fixed bug in database that caused nil objects to be returned by RemoveExpiredObjects

Added file containing test objects.
