+++ peer/send_test.go
@@ -6,9 +6,7 @@ package peer_test
 
 import (
 	"bytes"
-	"errors"
 	"net"
-	"sync"
 	"testing"
 	"time"
 
@@ -18,176 +16,6 @@ import (
 	"github.com/monetas/bmutil"
 	"github.com/monetas/bmutil/wire"
 )
-
-// MockConnection implements the Connection interface and is used to test
-// send without connecting to the real internet.
-type MockConnection struct {
-	closed      bool // Whether the connection has been closed.
-	connected   bool // Whether the connection is connected.
-	failure     bool // When this is true, sending or receiving messages returns an error.
-	connectFail bool // When this is true, the connection cannot connect.
-	done        chan struct{}
-	failChan    chan struct{}
-	reply       chan wire.Message
-	send        chan wire.Message
-	addr        net.Addr
-	mutex       sync.RWMutex
-	failmtx     sync.Mutex
-}
-
-func (mock *MockConnection) WriteMessage(msg wire.Message) error {
-	if mock.closed {
-		return errors.New("Connection closed.")
-	}
-
-	if mock.failure {
-		return errors.New("Mock Connection set to fail.")
-	}
-
-	mock.reply <- msg
-
-	return nil
-}
-
-// MockRead allows for tests as to whether a message has been sent or not.
-// It accepts an extra channel that can be read from if no message has been
-// sent. Normally, Read blocks until a message is received.
-func (mock *MockConnection) MockRead(reset chan struct{}) wire.Message {
-	if mock.closed {
-		return nil
-	}
-
-	select {
-	case <-mock.done:
-		return nil
-	case message := <-mock.reply:
-		return message
-	case <-reset:
-		return nil
-	}
-}
-
-func (mock *MockConnection) ReadMessage() (wire.Message, error) {
-	mock.mutex.RLock()
-	fail := mock.failure
-	mock.mutex.RUnlock()
-
-	if fail {
-		return nil, errors.New("Mock Connection set to fail.")
-	}
-
-	mock.failmtx.Lock()
-	defer mock.failmtx.Unlock()
-
-	select {
-	case msg := <-mock.send:
-		return msg, nil
-	case <-mock.failChan:
-		return nil, errors.New("Mock Connection set to fail.")
-	}
-}
-
-func (mock *MockConnection) MockWrite(msg wire.Message) {
-	mock.send <- msg
-}
-
-func (mock *MockConnection) BytesWritten() uint64 {
-	return 0
-}
-
-func (mock *MockConnection) BytesRead() uint64 {
-	return 0
-}
-
-func (mock *MockConnection) LastWrite() time.Time {
-	return time.Time{}
-}
-
-func (mock *MockConnection) LastRead() time.Time {
-	return time.Time{}
-}
-
-func (mock *MockConnection) RemoteAddr() net.Addr {
-	return mock.addr
-}
-
-func (mock *MockConnection) LocalAddr() net.Addr {
-	return nil
-}
-
-func (mock *MockConnection) Close() {
-	if mock.closed {
-		return
-	}
-	mock.closed = true
-	mock.mutex.Lock()
-	mock.connected = false
-	mock.mutex.Unlock()
-	close(mock.done)
-
-	// Drain any incoming messages.
-close:
-	for {
-		select {
-		case <-mock.reply:
-		default:
-			break close
-		}
-	}
-}
-
-func (mock *MockConnection) Connected() bool {
-	mock.mutex.RLock()
-	defer mock.mutex.RUnlock()
-
-	return mock.connected
-}
-
-func (mock *MockConnection) Connect() error {
-	if mock.connectFail {
-		return errors.New("Connection set to fail.")
-	}
-	mock.mutex.Lock()
-	mock.connected = true
-	mock.mutex.Unlock()
-	return nil
-}
-
-func (mock *MockConnection) SetFailure(b bool) {
-	mock.mutex.Lock()
-	mock.failure = b
-	mock.mutex.Unlock()
-
-	if b == true {
-		// Drain any messages being sent now so that WriteMessage will
-		// immediately return an error.
-		select {
-		case <-mock.done:
-		default:
-		}
-
-		// Close the failure channel to ensure any messages being read
-		// will return an error.
-		close(mock.failChan)
-
-	} else {
-		mock.failmtx.Lock()
-		mock.failChan = make(chan struct{})
-		mock.failmtx.Unlock()
-	}
-}
-
-func NewMockConnection(addr net.Addr, connected bool, fails bool) *MockConnection {
-	return &MockConnection{
-		done:        make(chan struct{}),
-		reply:       make(chan wire.Message),
-		send:        make(chan wire.Message),
-		failChan:    make(chan struct{}),
-		connected:   connected,
-		connectFail: fails,
-		addr:        addr,
-	}
-}
 
 var mockAddr net.Addr = &net.TCPAddr{IP: net.ParseIP("192.168.0.1"), Port: 8333}
 
@@ -356,25 +184,6 @@ func TestRequestData(t *testing.T) {
 
 	//Start the queue again to make sure it shuts down properly before the test ends.
 	queue.Start(conn)
-
-	// Engineer a situation in which the data channel gets filled up
-	// and must be cleaned out.
-	i := 0
-	for {
-		objMsg := wire.NewMsgUnknownObject(555, time.Now(), wire.ObjectType(4), 1, 1, []byte{0, 0, 0, byte(i)})
-
-		msg, _ := wire.ToMsgObject(objMsg)
-		db.InsertObject(msg)
-		hashes := []*wire.InvVect{&wire.InvVect{Hash: *msg.InventoryHash()}}
-
-		err = queue.QueueDataRequest(hashes)
-		// Don't need to fill up the channel this time because there is
-		// no error condition to recreate.
-		if i == 10 {
-			break
-		}
-		i++
-	}
 
 	queue.Stop()
 }
